requirements.txt
pandas
numpy
scikit-learn
sqlalchemy
pymysql
joblib
pyyaml
flask
notebook
python-dotenv
streamlit